Splitting integration tests
